Moved table header and add button from editable-data-table
