+++ indigo-frontend/src/app/pages/experiment/stoichiometry/editable-data-table/editable-data-table.component.ts
@@ -2,7 +2,6 @@ import {
   Component,
   inject,
   input,
-  output,
   signal,
   ViewChild,
 } from '@angular/core';
@@ -28,7 +27,6 @@ import {
 import { MatInput } from '@angular/material/input';
 import { MatDivider } from '@angular/material/divider';
 import { FormsModule } from '@angular/forms';
-import { ButtonComponent } from '@core/components/common/button/button.component';
 import { DictionaryItemRef } from '@core/types/entities/dictionary.i';
 import {
   ColumnConfig,
@@ -37,12 +35,7 @@ import {
   FieldValue,
   UnitFieldValue,
 } from '../shared/editable-table.types';
-import { DropdownMenuItem } from '@/core/components/common/dropdown-menu/dropdown-menu.i';
-import { SelectComponent } from '@/core/components/common/select/select.component';
 import { ExperimentDetailService } from '@/core/services/experiment/experiment-detail.service';
-import { SIGNIFICANT_FIGURES } from '../significant-figures.constants';
-import { catchError } from 'rxjs/internal/operators/catchError';
-import { EMPTY } from 'rxjs/internal/observable/empty';
 import { MatIconButton } from '@angular/material/button';
 import { MatTooltip } from '@angular/material/tooltip';
 
@@ -69,8 +62,6 @@ import { MatTooltip } from '@angular/material/tooltip';
     MatDivider,
     FormsModule,
     CommonModule,
-    ButtonComponent,
-    SelectComponent,
     MatIconButton,
     MatTooltip,
   ],
@@ -82,18 +73,13 @@ export class EditableDataTableComponent<TRow = unknown> {
   readonly ColumnInputType = ColumnInputType;
   @ViewChild(MatTable) table?: MatTable<TRow>;
 
-  title = input.required<string>();
   dataSource = input.required<TRow[] | null>();
   columns = input.required<ColumnConfig<TRow>[]>();
   displayedColumns = input.required<string[]>();
   emptyMessage = input<string>('No data available');
   loadingMessage = input<string>('Loading...');
-  showAddButton = input<boolean>(true);
-  showSignificantFigures = input<boolean>(false);
   expandableConfig = input<ExpandableConfig<TRow> | null>(null);
-  readonly items = signal<DropdownMenuItem[]>([...SIGNIFICANT_FIGURES]);
 
-  addRow = output<void>();
 
   expandedRows = signal<Set<TRow>>(new Set());
 
@@ -116,9 +102,6 @@ export class EditableDataTableComponent<TRow = unknown> {
       : null;
   }
 
-  onAddRow() {
-    this.addRow.emit();
-  }
 
   toggleRow(row: TRow) {
     const expanded = this.expandedRows();
@@ -142,28 +125,5 @@ export class EditableDataTableComponent<TRow = unknown> {
     }
     return this.displayedColumns();
   }
-
-  onSignificantFiguresChange(value: string | string[] | null): void {
-    const parsedValue = this.parseSignificantFigure(value);
-    if (parsedValue === null) return;
-
-    this.experimentDetailService
-      .updateDataModel({
-        type: 'SetExperimentSignificantFigures',
-        significantFigures: parsedValue,
-      })
-      .pipe(catchError(() => EMPTY))
-      .subscribe();
-  }
-
-  detailRow = (_index: number, row: TRow) => this.isRowExpanded(row);
-
-  private parseSignificantFigure(
-    value: string | string[] | null,
-  ): number | null {
-    if (!value || Array.isArray(value)) return null;
-
-    const parsed = Number.parseInt(value, 10);
-    return Number.isNaN(parsed) ? null : parsed;
-  }
+    detailRow = (_index: number, row: TRow) => this.isRowExpanded(row);
 }
